XMLHttpRequest/open-method-insecure.htm

<!DOCTYPE html>
<html>
  <head>
    <title>XMLHttpRequest: open() - "insecure" methods</title>
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
  </head>
  <body>
    <div id="log"></div>
    <script>
      function method(method) {
        test(function() {
          var client = new XMLHttpRequest()
          assert_throws("SECURITY_ERR", function() { client.open(method, "...") })
        }, document.title + " (" + method + ")")
      }
      method("track")
      method("TRACK")
      method("trAck")
      method("TRACE")
      method("trace")
      method("traCE")
      method("connect")
      method("CONNECT")
      method("connECT")
    </script>
  </body>
</html>
